requirements.txt
# Flask Web Framework
Flask==2.3.0
flask-socketio==5.3.0

# WebSocket support
python-socketio==5.9.0
python-engineio==4.7.0

# Raspberry Pi GPIO
RPi.GPIO==0.7.1

# I2C communication for LCD
smbus2==0.4.2

# SPI communication for Dot Matrix
spidev==3.6

# Web server
eventlet==0.33.3
